+++ packages/system/src/foundry/data-models/register.ts
@@ -9,11 +9,14 @@ import {
   WeaponDataModel,
 } from "./item-types";
 import { SkillDataModel } from "./skill";
+import { StarshipDataModel, VehicleDataModel } from "./machine";
 
 export function registerD6System2eDataModels(): void {
   CONFIG.Actor.dataModels.character = CharacterDataModel;
   CONFIG.Actor.dataModels.creature = CharacterDataModel;
   CONFIG.Actor.dataModels.npc = CharacterDataModel;
+  CONFIG.Actor.dataModels.starship = StarshipDataModel;
+  CONFIG.Actor.dataModels.vehicle = VehicleDataModel;
   CONFIG.Item.dataModels.skill = SkillDataModel;
   CONFIG.Item.dataModels.specialization = SpecializationDataModel;
   CONFIG.Item.dataModels.advantage = AdvantageDataModel;